+++ app/models/contact.rb
@@ -22,8 +22,6 @@ class Contact < ApplicationRecord
   alias_attribute :kind, :ident_type
   alias_attribute :copy_from_id, :original_id # Old attribute name; for PaperTrail
 
-  accepts_nested_attributes_for :legal_documents
-
   scope :email_verification_failed, lambda {
     joins('LEFT JOIN email_address_verifications emv ON contacts.email = emv.email')
       .where('success = false and verified_at IS NOT NULL')