+++ db/migrate/20191219112434_change_legal_documents_path_to_not_null.rb
@@ -1,0 +1,5 @@
+class ChangeLegalDocumentsPathToNotNull < ActiveRecord::Migration[5.0]
+  def change
+    change_column_null :legal_documents, :path, false
+  end
+end