+++ db/migrate/20191219124429_change_legal_documents_document_type_to_not_null.rb
@@ -1,0 +1,5 @@
+class ChangeLegalDocumentsDocumentTypeToNotNull < ActiveRecord::Migration[5.0]
+  def change
+    change_column_null :legal_documents, :document_type, false
+  end
+end